requirements.txt
appdirs==1.4.4
Arpeggio==2.0.3
beautifulsoup4==4.14.3
bitarray==3.8.0
bitmath==1.3.3.1
bitstring==4.3.1
blessed==1.17.12
bpylist2==4.1.1
certifi==2025.11.12
cffi==2.0.0
cgmetadata==0.2.0
charset-normalizer==3.4.4
click==8.3.1
idna==3.11
jedi==0.19.2
mac-alias==2.2.2
makelive==0.6.2
Mako==1.2.4
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==8.14.0
objexplore==1.6.3
osxmetadata==1.4.0
osxphotos==0.74.1
packaging==25.0
parso==0.8.5
pathvalidate==3.3.1
photoscript==0.5.2
prompt_toolkit==3.0.52
ptpython==3.0.32
py-applescript==1.0.3
pycparser==2.23
Pygments==2.19.2
pyobjc-core==12.1
pyobjc-framework-AppleScriptKit==12.1
pyobjc-framework-AppleScriptObjC==12.1
pyobjc-framework-AVFoundation==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Contacts==12.1
pyobjc-framework-CoreAudio==12.1
pyobjc-framework-CoreLocation==12.1
pyobjc-framework-CoreMedia==12.1
pyobjc-framework-CoreML==12.1
pyobjc-framework-CoreServices==12.1
pyobjc-framework-FSEvents==12.1
pyobjc-framework-Metal==12.1
pyobjc-framework-Photos==12.1
pyobjc-framework-Quartz==12.1
pyobjc-framework-ScriptingBridge==12.1
pyobjc-framework-UniformTypeIdentifiers==12.1
pyobjc-framework-Vision==12.1
python-dotenv==1.2.1
pytimeparse2==1.7.1
PyYAML==6.0.3
requests==2.32.5
rich==13.9.4
rich-theme-manager==0.11.0
shortuuid==1.0.13
six==1.17.0
soupsieve==2.8
strpdatetime==0.4.1
tenacity==8.5.0
textX==4.3.0
toml==0.10.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
utitools==0.4.0
wcwidth==0.2.14
wheel==0.45.1
whenever==0.8.10
wrapt==1.17.3
wurlitzer==3.1.1
xattr==1.3.0
xdg-base-dirs==6.0.2
